issue #6: add support for downloading without deletion
This change adds good support for downloading without deletection, while still
allowing for polling to continue to know that it does not need to download this
file again in the future, even if the file is moved out of the downloads directory
(as I feel may often be the case).  Removing the '-k' will remove these files from
put.io without downloading again.